Back_end/requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
Bitdesc==0.1.4
blinker==1.7.0
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.2.0
confection==0.1.4
contourpy==1.2.0
cors==1.0.1
cryptography==41.0.7
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
distro==1.8.0
faiss-cpu==1.7.4
Faker==20.1.0
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flatbuffers==23.5.26
fonttools==4.45.1
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl#sha256=37e9f1f6a278a5138fabdabcc92cc559da917f9b24c76f0adf6758720d7eab10
frozenlist==1.4.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
imageio==2.33.0
implicit==0.7.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.1
langcodes==3.3.0
langsmith==0.0.71
lazy_loader==0.3
mahotas==1.4.13
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mediapipe==0.10.9
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
openai==1.5.0
opencv-contrib-python==4.9.0.80
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==3.11.0
preshed==3.0.9
protobuf==3.20.3
PyAudio==0.2.14
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
selenium==4.15.2
shapely==2.0.2
six==1.16.0
sklearn==0.0.post11
smart-open==6.4.0
sniffio==1.3.0
snowflake-connector-python==3.5.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
SQLAlchemy==2.0.23
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.5.2
tldextract==5.1.1
tomlkit==0.12.3
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
